requirements.txt
absl-py==2.2.2
accelerate==1.2.1
aiosignal==1.3.2
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==25.3.0
av==12.3.0
blessings==1.7
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.0
cloudpickle==3.1.1
coacd==1.0.7
colorama==0.4.6
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
decord==0.6.0
Deprecated==1.2.18
diffusers==0.30.2
dill==0.4.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
fastparquet==2024.11.0
filelock @ file:///croot/filelock_1744281381737/work
flatbuffers==25.2.10
fonttools==4.58.0
freetype-py==2.5.1
frozenlist==1.6.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gast==0.6.0
genesis-world @ git+https://github.com/Genesis-Embodied-AI/Genesis.git@97732afcff357ebe5fe69cf50366758d4cb6fe3a
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
-e git+ssh://git@github.com/chuanEgg/Isaac-GR00T.git@c93ea601ac6c5c11f66eed2833017e0f68e99dc4#egg=gr00t
grpcio==1.71.0
gymnasium==1.0.0
h5py==3.12.1
huggingface-hub==0.31.2
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1741710844255/work
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==2.15.0
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.9
lazy_loader==0.4
libclang==18.1.1
libigl==2.5.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml-dtypes==0.2.0
moviepy==2.1.2
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
mujoco==3.3.2
mypy_extensions==1.1.0
networkx @ file:///croot/networkx_1737039604450/work
numba==0.61.2
numpy==1.26.4
numpydantic==1.6.7
oauthlib==3.2.2
omegaconf==2.3.0
ompl==1.7.0
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
OpenEXR==3.3.3
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
peft==0.14.0
pettingzoo==1.25.0
pillow==10.4.0
pipablepytorch3d==0.7.6
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
proglog==0.1.12
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycollada==0.9
pydantic==2.7.1
pydantic_core==2.18.2
PyGEL3D==0.6.1
pyglet==2.1.6
pygltflib==1.16.0
Pygments==2.19.1
pymeshlab==2023.12.post3
PyOpenGL==3.1.9
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyvista==0.45.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
scooby==0.10.1
sentry-sdk==2.28.0
setproctitle==1.3.6
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
taichi==1.7.3
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tetgen==0.6.5
tianshou==0.5.1
tifffile==2025.5.10
timm==1.0.14
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.45.2
trimesh==4.6.9
triton==3.1.0
typeguard==4.4.2
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tyro==0.9.17
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
vtk==9.4.2
wandb==0.18.0
Werkzeug==3.1.3
wrapt==1.14.1
yacs==0.1.8
z3-solver==4.15.0.0
zipp==3.21.0
zmq==0.0.0
